Fully abstract dependent logic out from main file
